requirements.txt
biopython==1.70
Django==2.0.1
obonet==0.2.2
psycopg2==2.7.3.2
pysam==0.13
tqdm==4.19.5
typing==3.6.4
